feat(ab test): disable related tree items when operation in progress to avoid concurrency error (#2311)
