+++ packages/vscode-extension/src/commonlib/azureLogin.ts
@@ -45,7 +45,7 @@ import {
   TelemetryErrorType,
 } from "../telemetry/extTelemetryEvents";
 import { VS_CODE_UI } from "../extension";
-import TreeViewManagerInstance from "../commandsTreeViewProvider";
+import TreeViewManagerInstance from "../treeview/treeViewManager";
 import * as path from "path";
 import * as fs from "fs-extra";
 import * as commonUtils from "../debug/commonUtils";
@@ -141,8 +141,9 @@ export class AzureAccountManager extends login implements AzureAccountProvider {
   }
 
   private isUserLogin(): boolean {
-    const azureAccount: AzureAccount =
-      vscode.extensions.getExtension<AzureAccount>("ms-vscode.azure-account")!.exports;
+    const azureAccount: AzureAccount = vscode.extensions.getExtension<AzureAccount>(
+      "ms-vscode.azure-account"
+    )!.exports;
     return azureAccount.status === "LoggedIn";
   }
 
@@ -167,8 +168,9 @@ export class AzureAccountManager extends login implements AzureAccountProvider {
 
   private doGetAccountCredentialAsync(): Promise<TokenCredentialsBase | undefined> {
     if (this.isUserLogin()) {
-      const azureAccount: AzureAccount =
-        vscode.extensions.getExtension<AzureAccount>("ms-vscode.azure-account")!.exports;
+      const azureAccount: AzureAccount = vscode.extensions.getExtension<AzureAccount>(
+        "ms-vscode.azure-account"
+      )!.exports;
       // Choose one tenant credential when users have multi tenants. (TODO, need to optize after UX design)
       // 1. When azure-account-extension has at least one subscription, return the first one credential.
       // 2. When azure-account-extension has no subscription and has at at least one session, return the first session credential.
@@ -306,8 +308,9 @@ export class AzureAccountManager extends login implements AzureAccountProvider {
    */
   async listSubscriptions(): Promise<SubscriptionInfo[]> {
     await this.getAccountCredentialAsync();
-    const azureAccount: AzureAccount =
-      vscode.extensions.getExtension<AzureAccount>("ms-vscode.azure-account")!.exports;
+    const azureAccount: AzureAccount = vscode.extensions.getExtension<AzureAccount>(
+      "ms-vscode.azure-account"
+    )!.exports;
     const arr: SubscriptionInfo[] = [];
     if (azureAccount.status === "LoggedIn") {
       if (azureAccount.subscriptions.length > 0) {
@@ -329,8 +332,9 @@ export class AzureAccountManager extends login implements AzureAccountProvider {
    */
   async setSubscription(subscriptionId: string): Promise<void> {
     if (this.isUserLogin()) {
-      const azureAccount: AzureAccount =
-        vscode.extensions.getExtension<AzureAccount>("ms-vscode.azure-account")!.exports;
+      const azureAccount: AzureAccount = vscode.extensions.getExtension<AzureAccount>(
+        "ms-vscode.azure-account"
+      )!.exports;
       for (let i = 0; i < azureAccount.subscriptions.length; ++i) {
         const item = azureAccount.subscriptions[i];
         if (item.subscription.subscriptionId == subscriptionId) {
@@ -366,8 +370,9 @@ export class AzureAccountManager extends login implements AzureAccountProvider {
   }
 
   getAzureAccount(): AzureAccount {
-    const azureAccount: AzureAccount =
-      vscode.extensions.getExtension<AzureAccount>("ms-vscode.azure-account")!.exports;
+    const azureAccount: AzureAccount = vscode.extensions.getExtension<AzureAccount>(
+      "ms-vscode.azure-account"
+    )!.exports;
     return azureAccount;
   }
 
@@ -390,8 +395,9 @@ export class AzureAccountManager extends login implements AzureAccountProvider {
   }
 
   async addStatusChangeEvent() {
-    const azureAccount: AzureAccount =
-      vscode.extensions.getExtension<AzureAccount>("ms-vscode.azure-account")!.exports;
+    const azureAccount: AzureAccount = vscode.extensions.getExtension<AzureAccount>(
+      "ms-vscode.azure-account"
+    )!.exports;
     AzureAccountManager.currentStatus = azureAccount.status;
     if (AzureAccountManager.currentStatus === "LoggedIn") {
       const subscriptioninfo = await this.readSubscription();
@@ -428,7 +434,7 @@ export class AzureAccountManager extends login implements AzureAccountProvider {
   getAccountInfo(): Record<string, string> | undefined {
     const azureAccount = this.getAzureAccount();
     if (azureAccount.status === loggedIn) {
-      return this.getJsonObject() as unknown as Record<string, string>;
+      return (this.getJsonObject() as unknown) as Record<string, string>;
     } else {
       return undefined;
     }
