+++ packages/vscode-extension/src/handlers.ts
@@ -53,8 +53,8 @@ import AzureAccountManager from "./commonlib/azureLogin";
 import AppStudioTokenInstance from "./commonlib/appStudioLogin";
 import AppStudioCodeSpaceTokenInstance from "./commonlib/appStudioCodeSpaceLogin";
 import VsCodeLogInstance from "./commonlib/log";
-import { TreeViewCommand } from "./commandsTreeViewProvider";
-import TreeViewManagerInstance from "./commandsTreeViewProvider";
+import { TreeViewCommand } from "./treeview/commandsTreeViewProvider";
+import TreeViewManagerInstance from "./treeview/treeViewManager";
 import { ExtTelemetry } from "./telemetry/extTelemetry";
 import {
   TelemetryEvent,
@@ -95,6 +95,8 @@ import { TreatmentVariables } from "./exp/treatmentVariables";
 import { StringContext } from "./utils/stringContext";
 import { ext } from "./extensionVariables";
 import { InputConfigsFolderName } from "@microsoft/teamsfx-api";
+import { CoreCallbackEvent } from "@microsoft/teamsfx-api";
+import { CommandsWebviewProvider } from "./treeview/commandsWebviewProvider";
 import { LocalEnvironment } from "./constants";
 
 export let core: FxCore;
@@ -175,6 +177,7 @@ export async function activate(): Promise<Result<Void, FxError>> {
       ui: VS_CODE_UI,
     };
     core = new FxCore(tools);
+    registerCoreEvents();
     await registerAccountTreeHandler();
     await registerEnvTreeHandler();
     await openMarkdownHandler();
@@ -191,6 +194,36 @@ export async function activate(): Promise<Result<Void, FxError>> {
     return err(FxError);
   }
   return result;
+}
+
+function registerCoreEvents() {
+  const developmentView = TreeViewManagerInstance.getTreeView("teamsfx-development");
+  if (developmentView instanceof CommandsWebviewProvider) {
+    core.on(CoreCallbackEvent.lock, () => {
+      (TreeViewManagerInstance.getTreeView(
+        "teamsfx-development"
+      ) as CommandsWebviewProvider).onLockChanged(true);
+    });
+    core.on(CoreCallbackEvent.unlock, () => {
+      (TreeViewManagerInstance.getTreeView(
+        "teamsfx-development"
+      ) as CommandsWebviewProvider).onLockChanged(false);
+    });
+  }
+
+  const deploymentView = TreeViewManagerInstance.getTreeView("teamsfx-deployment");
+  if (deploymentView instanceof CommandsWebviewProvider) {
+    core.on(CoreCallbackEvent.lock, () => {
+      (TreeViewManagerInstance.getTreeView(
+        "teamsfx-deployment"
+      ) as CommandsWebviewProvider).onLockChanged(true);
+    });
+    core.on(CoreCallbackEvent.unlock, () => {
+      (TreeViewManagerInstance.getTreeView(
+        "teamsfx-deployment"
+      ) as CommandsWebviewProvider).onLockChanged(false);
+    });
+  }
 }
 
 export async function getAzureSolutionSettings(): Promise<AzureSolutionSettings | undefined> {
