+++ packages/vscode-extension/tsconfig.json
@@ -16,5 +16,5 @@
     // "noFallthroughCasesInSwitch": true, /* Report errors for fallthrough cases in switch statement. */
     // "noUnusedParameters": true,  /* Report errors on unused parameters. */
   },
-  "exclude": ["node_modules", ".vscode-test", "test-folder"]
+  "exclude": ["node_modules", ".vscode-test", "test-folder", "out"]
 }
